requirements.txt
Pillow==8.1.0
Markdown==3.3.4
numpy==1.21.4
plotly==4.14.3
pandas==1.2.0
streamlit==1.4.0
openpyxl==3.0.6
openpyxl==3.0.6
pyxlsb==1.0.9
XlsxWriter==3.0.2
tk==0.1.0
gunicorn==20.1.0